Refactor the no image page
